Move _find_kedro_project to utils
