+++ kedro/utils.py
@@ -3,7 +3,10 @@ of kedro package.
 """
 import importlib
 import os
-from typing import Any
+from pathlib import Path
+from typing import Any, Union
+
+_PYPROJECT = "pyproject.toml"
 
 
 def load_obj(obj_path: str, default_obj_path: str = "") -> Any:
@@ -29,3 +32,23 @@ def load_obj(obj_path: str, default_obj_path: str = "") -> Any:
 
 def _is_databricks() -> bool:
     return "DATABRICKS_RUNTIME_VERSION" in os.environ
+
+
+def _is_project(project_path: Union[str, Path]) -> bool:
+    metadata_file = Path(project_path).expanduser().resolve() / _PYPROJECT
+    if not metadata_file.is_file():
+        return False
+
+    try:
+        return "[tool.kedro]" in metadata_file.read_text(encoding="utf-8")
+    except Exception:  # noqa: broad-except
+        return False
+
+
+def _find_kedro_project(current_dir: Path) -> Any:  # pragma: no cover
+    while current_dir != current_dir.parent:
+        if _is_project(current_dir):
+            return current_dir
+        current_dir = current_dir.parent
+
+    return None
